refactor: move getMeetingDataCached into getMeetingData.ts
Consolidates all meeting data logic in one file, eliminating the
circular dependency between queries.ts and getMeetingData.ts.

Reuses getPeopleForCityCached and getPartiesForCityCached from
queries.ts instead of duplicating inline createCache calls.
